Switching types around in Polygon to play nicely with SDL.

+++ polygon.cc
@@ -2,13 +2,13 @@
 
 #include <cstring>
 
-Polygon::Polygon(size_t num, const uint32_t* x, const uint32_t* y, struct Colour colour) 
+Polygon::Polygon(size_t num, const int16_t* x, const int16_t* y, struct Colour colour) 
 	: _num(num), _colour(colour) {
-	_x = new uint32_t[_num];
-	_y = new uint32_t[_num];
+	_x = new int16_t[_num];
+	_y = new int16_t[_num];
 
-	::memcpy(_x, x, _num * sizeof(uint32_t));
-	::memcpy(_y, y, _num * sizeof(uint32_t));
+	::memcpy(_x, x, _num * sizeof(int16_t));
+	::memcpy(_y, y, _num * sizeof(int16_t));
 }
 
 Polygon::Polygon(const Polygon& other) 
@@ -25,14 +25,14 @@ Polygon& Polygon::operator=(const Polygon& rhs) {
 			delete[] _x;
 			delete[] _y;
 		}
-		_x = new uint32_t[rhs.num()];
-		_y = new uint32_t[rhs.num()];
+		_x = new int16_t[rhs.num()];
+		_y = new int16_t[rhs.num()];
 		_cap = rhs.num();
 	}
 
 	_num = rhs.num();
-	::memcpy(_x, rhs.x(), _num * sizeof(uint32_t));
-	::memcpy(_y, rhs.y(), _num * sizeof(uint32_t));
+	::memcpy(_x, rhs.x(), _num * sizeof(int16_t));
+	::memcpy(_y, rhs.y(), _num * sizeof(int16_t));
 
 	_colour = rhs.colour();
 
@@ -43,11 +43,11 @@ size_t Polygon::num() const {
 	return _num;
 }
 
-const uint32_t* Polygon::x() const {
+const int16_t* Polygon::x() const {
 	return _x;
 }
 
-const uint32_t* Polygon::y() const {
+const int16_t* Polygon::y() const {
 	return _y;
 }
